pick up binary name from the main script
